front: split NGE-to-OSRD node-related logic in separate file
