+++ front/src/applications/operationalStudies/views/Scenario/components/MacroEditor/ngeToOsrd/ngeToOsrd.ts
@@ -8,7 +8,6 @@ import {
   osrdEditoastApi,
   type MacroNodeForm,
   type PacedTrain,
-  type SearchResultItemOperationalPoint,
   type TrainSchedule,
   type PathItemLocation,
 } from 'common/api/osrdEditoastApi';
@@ -39,22 +38,7 @@ import {
   isTrainScheduleId,
 } from 'utils/trainId';
 
-import {
-  DEFAULT_PACED_TRAIN_PAYLOAD,
-  DEFAULT_TRAIN_SCHEDULE_PAYLOAD,
-  DEFAULT_TIME_WINDOW,
-  TRAINRUN_DIRECTIONS,
-} from './consts';
-import MacroEditorState from './MacroEditorState';
-import type { NodeIndexed } from './MacroEditorState';
-import {
-  createMacroNode,
-  deleteMacroNodeByNgeId,
-  getFrequencyFromFrequencyId,
-  getTrainCategoryFromTrainrunCategoryId,
-  updateMacroNode,
-} from './utils';
-import { checkChangeGroups } from '../ManageTimetableItem/helpers/buildPacedTrainException';
+import { checkChangeGroups } from '../../ManageTimetableItem/helpers/buildPacedTrainException';
 import type {
   FreeFloatingTextDto,
   NetzgrafikDto,
@@ -64,7 +48,20 @@ import type {
   TimeLockDto,
   TrainrunDto,
   LabelDto,
-} from '../NGE/types';
+} from '../../NGE/types';
+import {
+  DEFAULT_PACED_TRAIN_PAYLOAD,
+  DEFAULT_TRAIN_SCHEDULE_PAYLOAD,
+  DEFAULT_TIME_WINDOW,
+  TRAINRUN_DIRECTIONS,
+} from '../consts';
+import MacroEditorState from '../MacroEditorState';
+import {
+  fetchStationSecondaryCode,
+  getFrequencyFromFrequencyId,
+  getTrainCategoryFromTrainrunCategoryId,
+} from '../utils';
+import { castNgeNode, handleNodeOperation } from './node';
 
 const getNodeById = (nodes: NodeDto[], nodeId: number | string) =>
   nodes.find((node) => node.id === nodeId);
@@ -361,25 +358,6 @@ const generatePathAndSchedule = (
   const path = generatePath(sections, nodes, trainrunDirection, state);
   const schedule = generateSchedule(sections, nodes, startDate, trainrunDirection);
   return { start_time: startDate.toISOString(), path, schedule };
-};
-
-const fetchStationSecondaryCode = async (
-  trigram: string,
-  infraId: number,
-  dispatch: AppDispatch
-) => {
-  const searchPayload = {
-    object: 'operationalpoint',
-    query: ['and', ['=', ['infra_id'], infraId], ['=', ['trigram'], trigram]],
-  };
-  const searchResults = (await dispatch(
-    osrdEditoastApi.endpoints.postSearch.initiate({
-      searchPayload,
-    })
-  ).unwrap()) as SearchResultItemOperationalPoint[];
-
-  const stationOp = searchResults.find((op) => op.ch === 'BV' || op.ch === '00');
-  return stationOp?.ch;
 };
 
 // TODO: drop this function once this PR is merged:
@@ -930,104 +908,6 @@ const handleNoteOperation = async ({
     }
     case 'delete': {
       await deleteMacroNote(state, dispatch, note.id);
-      break;
-    }
-    default:
-      break;
-  }
-};
-
-/**
- * Cast a NGE node to a node.
- */
-const castNgeNode = (
-  node: NetzgrafikDto['nodes'][0],
-  labels: NetzgrafikDto['labels']
-): Omit<NodeIndexed, 'path_item_key' | 'dbId'> => ({
-  ngeId: node.id,
-  trigram: node.betriebspunktName,
-  full_name: node.fullName,
-  connection_time: node.connectionTime,
-  position_x: Math.round(node.positionX),
-  position_y: Math.round(node.positionY),
-  labels: node.labelIds
-    .map((id) => {
-      const ngeLabel = labels.find((e) => e.id === id);
-      if (ngeLabel) return ngeLabel.label;
-      return null;
-    })
-    .filter((n) => n !== null),
-});
-
-const handleNodeOperation = async ({
-  state,
-  type,
-  node,
-  netzgrafikDto,
-  dispatch,
-}: {
-  state: MacroEditorState;
-  type: NGEEvent['type'];
-  node: NodeDto;
-  netzgrafikDto: NetzgrafikDto;
-  dispatch: AppDispatch;
-}): Promise<void> => {
-  const indexNode = state.getNodeByNgeId(node.id);
-  switch (type) {
-    case 'create':
-    case 'update': {
-      if (indexNode) {
-        if (indexNode.dbId) {
-          // Update the key if trigram has changed and key is based on it
-          let nodeKey = indexNode.path_item_key;
-          let trigram = node.betriebspunktName;
-          if (nodeKey.startsWith('trigram:') && indexNode.trigram !== trigram) {
-            if (!trigram.includes('/')) {
-              const secondaryCode = await fetchStationSecondaryCode(
-                trigram,
-                state.infraId,
-                dispatch
-              );
-              if (secondaryCode) {
-                trigram = `${trigram}/${secondaryCode}`;
-              }
-            }
-            nodeKey = `trigram:${trigram}`;
-          }
-          await updateMacroNode(state, dispatch, {
-            ...indexNode,
-            ...castNgeNode(node, netzgrafikDto.labels),
-            trigram,
-            dbId: indexNode.dbId,
-            path_item_key: nodeKey,
-          });
-        } else {
-          const newNode = {
-            ...indexNode,
-            ...castNgeNode(node, netzgrafikDto.labels),
-          };
-          // Create the node
-          await createMacroNode(state, dispatch, newNode, node.id);
-        }
-      } else {
-        // It's an unknown node, we need to create it in the db
-        // We assume that `betriebspunktName` is a trigram
-        const key = `trigram:${node.betriebspunktName}`;
-        // Create the node
-        await createMacroNode(
-          state,
-          dispatch,
-          {
-            ...castNgeNode(node, netzgrafikDto.labels),
-            path_item_key: key,
-          },
-          node.id
-        );
-      }
-      break;
-    }
-    case 'delete': {
-      if (indexNode) await deleteMacroNodeByNgeId(state, dispatch, node.id);
       break;
     }
     default:
